+++ src/langserver/toolsLangServerClient.ts
@@ -63,7 +63,8 @@ export class ToolsLangServerClient implements vscode.Disposable {
     _childProcess : cp.ChildProcess | undefined;
     _connection : rpc.MessageConnection | undefined;
     _alExtensionPath : string;
-    _alExtensionVersion : Version;    
+    _alExtensionVersion : Version;
+    errorLogUri : vscode.Uri | undefined;
 
     constructor(context : vscode.ExtensionContext, alExtensionPath : string, alExtensionVersion: Version) {
         this._context = context;
@@ -71,6 +72,7 @@ export class ToolsLangServerClient implements vscode.Disposable {
         this._connection = undefined;
         this._alExtensionPath = alExtensionPath;
         this._alExtensionVersion = alExtensionVersion;
+        this.errorLogUri = undefined;
         this.initialize();
     }
 
@@ -88,15 +90,18 @@ export class ToolsLangServerClient implements vscode.Disposable {
             let langServerPath : string;
 
             //find binaries path
-            if (this._alExtensionVersion.major <= 1)
+            if (this._alExtensionVersion.major <= 1) {
                 langServerPath = this._context.asAbsolutePath("bin/netframeworknav2018/AZALDevToolsServer.NetFrameworkNav2018.exe");
-            else {
+                this.errorLogUri = vscode.Uri.file(this._context.asAbsolutePath("bin/netframeworknav2018/log.txt"));
+            } else {                
                 langServerPath = this._context.asAbsolutePath("bin/netcore/" + platform + "/AZALDevToolsServer.NetCore");
+                this.errorLogUri = vscode.Uri.file(this._context.asAbsolutePath("bin/netcore/" + platform + "/log.txt"));
                 if (platform == "darwin") {
                     let fs = require('fs');
                     fs.chmodSync(langServerPath, 0o755);
                 }
             }
+
 
             //start child process
             this._childProcess = cp.spawn(langServerPath, [this._alExtensionPath]);